Simplify hexagonal-architecture-eval code
