+++ patterns/hexagonal-architecture-eval/src/main/java/com/baeldung/pattern/hexagonal/eval/application/adapters/AddressBookRepositoryImpl.java
@@ -27,5 +27,4 @@ public class AddressBookRepositoryImpl implements AddressBookRepository {
 
         return contacts.get(nickname);
     }
-
 }
