--- patterns/hexagonal-architecture-eval/src/main/java/com/baeldung/pattern/hexagonal/eval/domain/model/ContactName.java
@@ -1,37 +1,0 @@
-package main.java.com.baeldung.pattern.hexagonal.eval.domain.model;
-
-public class ContactName {
-    String title;
-    String firstName;
-    String lastName;
-
-    public ContactName(String title, String firstName, String lastName) {
-        this.title = title;
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-}
