+++ patterns/hexagonal-architecture-eval/src/main/java/com/baeldung/pattern/hexagonal/eval/domain/ports/DiallerService.java
@@ -2,5 +2,5 @@ package main.java.com.baeldung.pattern.hexagonal.eval.domain.ports;
 
 public interface DiallerService {
 
-    boolean dial(String name);
+    public boolean dial(String nickname);
 }
